refactor(digitsd): extract shared wififallback subprocess helper
